backend/upload-material/requirements.txt
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-docx==1.1.0
